# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Lab Launcher API",
    "version": "0.7.0",
    "description": "Pre-1.0 JSON API for the bundled frontend, trusted plugins, and private-deployment automation. Mutating API routes require the session CSRF token in the X-CSRF-Token header, except /api/bootstrap and /api/auth/login. Automation may instead authenticate with an Admin-issued API token via Authorization: Bearer; token requests skip CSRF, act with the token's role, and are rejected on session/profile endpoints (/api/auth/*, /api/me*, /api/bootstrap*)."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "security": [
    { "sessionCookie": [] },
    { "bearerToken": [] }
  ],
  "paths": {
...
    "/plugins": { "get": { "tags": ["Plugins"], "summary": "List installed plugins", "parameters": [{ "name": "updates", "in": "query", "schema": { "type": "string", "enum": ["1"] }, "description": "Admins may request update discovery with updates=1." }], "responses": { "200": { "description": "Plugins", "content": { "application/json": { "schema": { "type": "object", "properties": { "plugins": { "type": "array", "items": { "$ref": "#/components/schemas/Plugin" } } } } } } }, "401": { "$ref": "#/components/responses/Error" } } } },
    "/plugins/enabled-sections": { "get": { "tags": ["Plugins"], "summary": "List enabled plugin dashboard sections", "security": [], "responses": { "200": { "description": "Plugin sections", "content": { "application/json": { "schema": { "type": "object", "properties": { "sections": { "type": "array", "items": { "$ref": "#/components/schemas/GenericObject" } } } } } } }, "401": { "$ref": "#/components/responses/Error" } } } },
    "/plugins/{id}/logs": { "get": { "tags": ["Plugins"], "summary": "List plugin logs", "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "description": "Plugin logs", "content": { "application/json": { "schema": { "type": "object", "properties": { "logs": { "type": "array", "items": { "$ref": "#/components/schemas/AuditLog" } } } } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/plugins/reload": { "post": { "tags": ["Plugins"], "summary": "Reload installed plugins", "security": [{ "sessionCookie": [], "csrfToken": [] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/plugin-sources/github/versions": { "get": { "tags": ["Plugins"], "summary": "Discover GitHub plugin versions", "parameters": [{ "name": "repo", "in": "query", "required": true, "schema": { "type": "string" } }], "responses": { "200": { "description": "Version list", "content": { "application/json": { "schema": { "type": "object", "properties": { "versions": { "type": "array", "items": { "$ref": "#/components/schemas/GenericObject" } } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/plugin-sources/updates": { "get": { "tags": ["Plugins"], "summary": "Check plugin updates", "responses": { "200": { "description": "Updates", "content": { "application/json": { "schema": { "type": "object", "properties": { "updates": { "type": "array", "items": { "$ref": "#/components/schemas/GenericObject" } } } } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/plugin-sources/local/status": { "get": { "tags": ["Plugins"], "summary": "Read local plugin install status", "responses": { "200": { "description": "Local plugin status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GenericObject" } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/plugins/install": { "post": { "tags": ["Plugins"], "summary": "Install a trusted GitHub plugin", "security": [{ "sessionCookie": [], "csrfToken": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PluginInstallRequest" } } } }, "responses": { "201": { "description": "Installed plugin", "content": { "application/json": { "schema": { "type": "object", "properties": { "plugin": { "$ref": "#/components/schemas/Plugin" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/plugins/install-local": { "post": { "tags": ["Plugins"], "summary": "Install a trusted local plugin", "security": [{ "sessionCookie": [], "csrfToken": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PluginLocalInstallRequest" } } } }, "responses": { "201": { "description": "Installed plugin", "content": { "application/json": { "schema": { "type": "object", "properties": { "plugin": { "$ref": "#/components/schemas/Plugin" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/plugins/{id}/update": { "post": { "tags": ["Plugins"], "summary": "Update a GitHub plugin", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PluginInstallRequest" } } } }, "responses": { "200": { "description": "Updated plugin", "content": { "application/json": { "schema": { "type": "object", "properties": { "plugin": { "$ref": "#/components/schemas/Plugin" }, "rolledBack": { "type": "boolean" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/plugins/{id}": {
      "patch": { "tags": ["Plugins"], "summary": "Enable or disable a plugin", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "enabled": { "type": "boolean" } } } } } }, "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } },
      "delete": { "tags": ["Plugins"], "summary": "Delete a plugin", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/plugins/{id}/config": { "put": { "tags": ["Plugins"], "summary": "Update scoped plugin config fields", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "config": { "$ref": "#/components/schemas/GenericObject" } } } } } }, "responses": { "200": { "description": "Config update result", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/Ok" }, { "type": "object", "properties": { "updatedFields": { "type": "array", "items": { "type": "string" } }, "rejectedFields": { "type": "array", "items": { "type": "string" } } } }] } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } } }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "hll.sid" },
      "csrfToken": { "type": "apiKey", "in": "header", "name": "X-CSRF-Token" },
      "bearerToken": { "type": "http", "scheme": "bearer", "description": "Admin-issued API token (hll_ prefix). Skips CSRF; not accepted on session/profile endpoints." }
    },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "Ok": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } },
...
      "Error": { "description": "Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Ok": { "type": "object", "required": ["ok"], "properties": { "ok": { "type": "boolean" } }, "additionalProperties": true },
...
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } }, "additionalProperties": true },
      "GenericObject": { "type": "object", "additionalProperties": true },
...
      "AuditLog": { "type": "object", "properties": { "id": { "type": "integer" }, "level": { "type": "string" }, "action": { "type": "string" }, "actorUserId": { "type": ["integer", "null"] }, "actorUsername": { "type": ["string", "null"] }, "ip": { "type": ["string", "null"] }, "details": { "$ref": "#/components/schemas/GenericObject" }, "createdAt": { "type": "string" } }, "additionalProperties": true },
...
      "Plugin": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "sourceUrl": { "type": "string" }, "sourceType": { "type": "string" }, "version": { "type": "string" }, "enabled": { "type": "boolean" }, "manifest": { "$ref": "#/components/schemas/GenericObject" }, "config": { "$ref": "#/components/schemas/GenericObject" }, "installedHash": { "type": "string" }, "lifecycle": { "type": "string" }, "lastError": { "type": ["string", "null"] } }, "additionalProperties": true },
      "PluginInstallRequest": { "type": "object", "required": ["repoUrl", "version", "trustConfirmed"], "properties": { "repoUrl": { "type": "string" }, "version": { "type": "string" }, "expectedSha256": { "type": "string" }, "trustConfirmed": { "type": "boolean", "const": true } }, "additionalProperties": false },
      "PluginLocalInstallRequest": { "type": "object", "required": ["path", "trustConfirmed"], "properties": { "path": { "type": "string" }, "trustConfirmed": { "type": "boolean", "const": true } }, "additionalProperties": false },
...
    }
  }
}